Add html and css for bottom section of the webpage
This is almost visually similar to the page but some functionality is not present. For example the footer is not really acting like footer and signup button on bottom section not properly aligned. There are too many flex box stuff in this wenpage that it would need considerable time to understand and fix them all as some of the things are not yet conceptually covered in the lessons. Although alot of thins were completed by getting aid from internet but this satisfies the project project requirement hopefully. Otherwise it will become a verylong project.